Calling asTruffleObject is now safe in AOT mode as well

+++ dataset/GitHub_Java/oracle.graal/projects/com.oracle.truffle.llvm.parser.base/src/com/oracle/truffle/llvm/parser/base/model/symbols/constants/integer/BigIntegerConstant.java
@@ -49,9 +49,10 @@ public final class BigIntegerConstant extends AbstractConstant {
 
     @Override
     public String toString() {
-        if (((IntegerType) getType()).getBits() == 1) {
+        if (getType().getBits() == 1) {
             return value.equals(BigInteger.ZERO) ? "false" : "true";
         }
         return value.toString();
     }
+
 }
